test: expose specification detail workflows observably (#622)
