+++ tests/unit/requirements-specification-detail-client.suite.tsx
@@ -1,5 +1,4 @@
 import {
-  act,
   fireEvent,
   render,
   screen,
@@ -22,17 +21,15 @@ import type {
   SpecificationListItem,
   SpecificationPreloadError,
 } from '@/lib/specifications/preload-types'
-import { SPECIFICATION_PRELOAD_ERROR_KEYS } from '@/lib/specifications/preload-types'
+import { registerAddingLocalTests } from './requirements-specification-detail-adding-local.suite'
+import { registerAvailabilityTests } from './requirements-specification-detail-availability.suite'
 import { registerGeneratedOutputTests } from './requirements-specification-detail-generated-output.suite'
+import { registerMetadataTableTests } from './requirements-specification-detail-metadata-table.suite'
 import { registerNeedsReferenceTests } from './requirements-specification-detail-needs-references.suite'
 import { registerPaginationTests } from './requirements-specification-detail-pagination.suite'
+import { registerResilienceTests } from './requirements-specification-detail-resilience.suite'
 import { registerSelectionTests } from './requirements-specification-detail-selection.suite'
 
-const requirementsTableMock = vi.fn()
-const lazyFeatureState = vi.hoisted(() => ({
-  aiRenderSpy: vi.fn(),
-  importRenderSpy: vi.fn(),
-}))
 const intlState = vi.hoisted(() => ({
   locale: 'en',
   selectionActionLimitExceeded: vi.fn(),
@@ -40,12 +37,6 @@ const intlState = vi.hoisted(() => ({
 }))
 const navigationState = vi.hoisted(() => ({
   searchParams: new URLSearchParams(),
-}))
-const requirementDetailState = vi.hoisted(() => ({
-  renderSpy: vi.fn(),
-}))
-const localRequirementDetailState = vi.hoisted(() => ({
-  renderSpy: vi.fn(),
 }))
 const pdfDownloadState = vi.hoisted(() => ({
   clearError: vi.fn(),
@@ -115,7 +106,6 @@ vi.mock('@/app/[locale]/requirements/[id]/requirement-detail-client', () => ({
     removeFromSpecificationDisabled?: boolean
     requirementId: number
   }) => {
-    requirementDetailState.renderSpy(props)
     return (
       <div>
         {`Requirement detail ${props.requirementId}`}
@@ -140,7 +130,6 @@ vi.mock('@/app/[locale]/requirements/[id]/requirement-detail-client', () => ({
 
 vi.mock('@/components/SpecificationLocalRequirementDetailClient', () => ({
   default: (props: { localRequirementId: number; onChange?: () => void }) => {
-    localRequirementDetailState.renderSpy(props)
     return (
       <button onClick={() => void props.onChange?.()} type="button">
         {`Local requirement detail ${props.localRequirementId}`}
@@ -155,7 +144,6 @@ vi.mock('@/components/LazyAiRequirementGenerator', () => ({
     open?: boolean
     returnFocusTarget?: HTMLElement | null
   }) => {
-    lazyFeatureState.aiRenderSpy(props)
     return props.open ? (
       <section aria-label="AI authoring">
         <button
@@ -175,13 +163,24 @@ vi.mock('@/components/LazyRequirementsImportDialog', () => ({
     open?: boolean
     returnFocusTarget?: HTMLElement | null
   }) => {
-    lazyFeatureState.importRenderSpy(props)
     return props.open ? (
       <section aria-label="Import review">
-        <button onClick={() => void props.onClose?.(false)} type="button">
+        <button
+          onClick={async () => {
+            await props.onClose?.(false)
+            props.returnFocusTarget?.focus()
+          }}
+          type="button"
+        >
           cancel import review
         </button>
-        <button onClick={() => void props.onClose?.(true)} type="button">
+        <button
+          onClick={async () => {
+            await props.onClose?.(true)
+            props.returnFocusTarget?.focus()
+          }}
+          type="button"
+        >
           finish import review
         </button>
       </section>
@@ -294,6 +293,7 @@ vi.mock('@/components/RequirementsTable', () => ({
       itemRef?: string
       needsReference?: string | null
       requirementPackageIds?: number[]
+      specificationItemStatusId?: number | null
     }[]
     renderExpanded?: (id: number) => ReactNode
     selectable?: boolean
@@ -306,7 +306,6 @@ vi.mock('@/components/RequirementsTable', () => ({
     stickyTitleActions?: ReactNode
     visibleColumns?: string[]
   }) => {
-    requirementsTableMock(props)
     const tableKind = props.excludeColumns?.includes('needsReference')
       ? 'available'
       : 'items'
@@ -319,12 +318,55 @@ vi.mock('@/components/RequirementsTable', () => ({
       renderedRows.length > 25 && lastRenderedRow
         ? [...renderedRows.slice(0, 24), lastRenderedRow]
         : renderedRows
+    const renderFloatingAction = (
+      action: NonNullable<typeof props.floatingActions>[number],
+    ) => (
+      <div key={action.id}>
+        <button
+          aria-label={action.ariaLabel}
+          data-developer-mode-context={action.developerModeContext}
+          data-developer-mode-name="table action"
+          data-developer-mode-value={action.developerModeValue}
+          onClick={action.onClick}
+          type="button"
+        >
+          {action.icon}
+        </button>
+        {action.menuItems ? (
+          <div aria-label={`${action.id} menu`} role="menu">
+            {action.menuItems.map(menuItem =>
+              menuItem.id.startsWith('separator-') ? (
+                <hr key={menuItem.id} />
+              ) : menuItem.href ? (
+                <a href={menuItem.href} key={menuItem.id} role="menuitem">
+                  {menuItem.label}
+                </a>
+              ) : (
+                <button
+                  aria-description={
+                    'description' in menuItem
+                      ? String(menuItem.description)
+                      : undefined
+                  }
+                  disabled={menuItem.disabled}
+                  key={menuItem.id}
+                  onClick={event => menuItem.onClick?.(event.currentTarget)}
+                  role="menuitem"
+                  title={
+                    'tooltip' in menuItem ? String(menuItem.tooltip) : undefined
+                  }
+                  type="button"
+                >
+                  {menuItem.label}
+                </button>
+              ),
+            )}
+          </div>
+        ) : null}
+      </div>
+    )
     return (
-      <div
-        data-floating-action-rail-placement={
-          props.floatingActionRailPlacement ?? 'fixed-right'
-        }
-      >
+      <section aria-label={`${tableKind} requirements surface`}>
         <table aria-label={`${tableKind} requirements`}>
           <tbody>
             {renderedRowsInMock.map(row => (
@@ -335,46 +377,32 @@ vi.mock('@/components/RequirementsTable', () => ({
             ))}
           </tbody>
         </table>
-        <div data-testid="requirements-table-sticky-title">
+        <fieldset aria-label={`${tableKind} requirements sticky title`}>
           {props.stickyTitle}
-        </div>
-        <div data-testid="requirements-table-sticky-title-actions">
+        </fieldset>
+        <fieldset aria-label={`${tableKind} requirements sticky title actions`}>
           {props.stickyTitleActions}
-        </div>
-        <div data-testid={`requirements-table-${tableKind}-status`}>
-          {props.statusRow}
-        </div>
+        </fieldset>
+        {props.statusRow ? (
+          <div aria-label={`${tableKind} requirements status`} role="status">
+            {props.statusRow}
+          </div>
+        ) : null}
         {props.floatingActions
-          ?.filter(action => !action.hidden)
-          .map(action => (
-            <div key={action.id}>
-              <button
-                aria-label={action.ariaLabel}
-                data-developer-mode-context={action.developerModeContext}
-                data-developer-mode-name="table action"
-                data-developer-mode-value={action.developerModeValue}
-                onClick={action.onClick}
-                type="button"
-              >
-                {action.icon}
-              </button>
-              {action.menuItems ? (
-                <div aria-label={`${action.id} menu`} role="menu">
-                  {action.menuItems.map(menuItem => (
-                    <button
-                      disabled={menuItem.disabled}
-                      key={menuItem.id}
-                      onClick={event => menuItem.onClick?.(event.currentTarget)}
-                      role="menuitem"
-                      type="button"
-                    >
-                      {menuItem.label}
-                    </button>
-                  ))}
-                </div>
-              ) : null}
-            </div>
-          ))}
+          ?.filter(
+            action => !action.hidden && action.position === 'beforeColumns',
+          )
+          .map(renderFloatingAction)}
+        {props.columnPickerPlacement === 'betweenActions' ? (
+          <button aria-label="common.columns" type="button">
+            common.columns
+          </button>
+        ) : null}
+        {props.floatingActions
+          ?.filter(
+            action => !action.hidden && action.position !== 'beforeColumns',
+          )
+          .map(renderFloatingAction)}
         {props.requirementPackages?.map(requirementPackage => {
           const current = props.filterValues?.requirementPackageIds ?? []
           const active = current.includes(requirementPackage.id)
@@ -572,7 +600,7 @@ vi.mock('@/components/RequirementsTable', () => ({
           </button>
         ) : null}
         {`rows:${props.rows.length}`}
-      </div>
+      </section>
     )
   },
 }))
@@ -839,30 +867,6 @@ function searchParamsFromPath(path: string): URLSearchParams {
   return new URLSearchParams(path.split('?')[1] ?? '')
 }
 
-function latestItemsTableProps() {
-  const calls = requirementsTableMock.mock.calls.map(([props]) => props)
-  const itemsTable = calls
-    .slice()
-    .reverse()
-    .find(
-      props =>
-        props.floatingActionRailPlacement === 'inline-top' &&
-        !props.excludeColumns?.includes('needsReference'),
-    )
-  expect(itemsTable).toBeDefined()
-  return itemsTable as NonNullable<typeof itemsTable>
-}
-
-function latestAvailableTableProps() {
-  const calls = requirementsTableMock.mock.calls.map(([props]) => props)
-  const availableTable = calls
-    .slice()
-    .reverse()
-    .find(props => props.excludeColumns?.includes('needsReference'))
-  expect(availableTable).toBeDefined()
-  return availableTable as NonNullable<typeof availableTable>
-}
-
 function selectRequirementRows(ids: number[]) {
   for (const id of ids) {
     fireEvent.click(screen.getByRole('button', { name: `select-row-${id}` }))
@@ -955,16 +959,12 @@ const workflowContext = {
   set itemStatusPatchOk(value) {
     itemStatusPatchOk = value
   },
-  latestAvailableTableProps,
-  latestItemsTableProps,
-  lazyFeatureState,
   get localRequirementPostOk() {
     return localRequirementPostOk
   },
   set localRequirementPostOk(value) {
     localRequirementPostOk = value
   },
-  localRequirementDetailState,
   navigationState,
   get needsReferenceMutationHandler() {
     return needsReferenceMutationHandler
@@ -993,9 +993,7 @@ const workflowContext = {
   okJson,
   pdfDownloadState,
   renderRequirementsSpecificationDetailClient,
-  requirementDetailState,
   requirementRowNames,
-  requirementsTableMock,
   searchParamsFromPath,
   selectRequirementRows,
   specificationApiPath,
@@ -1054,10 +1052,7 @@ describe('RequirementsSpecificationDetailClient', () => {
     intlState.locale = 'en'
     intlState.selectionActionLimitExceeded.mockReset()
     intlState.selectionStatus.mockReset()
-    requirementDetailState.renderSpy.mockReset()
-    localRequirementDetailState.renderSpy.mockReset()
     vi.mocked(useReducedMotion).mockReturnValue(false)
-    requirementsTableMock.mockReset()
     pdfDownloadState.clearError.mockReset()
     pdfDownloadState.download.mockReset()
     pdfDownloadState.download.mockResolvedValue(undefined)
@@ -1560,2424 +1555,11 @@ describe('RequirementsSpecificationDetailClient', () => {
 
   registerGeneratedOutputTests(workflowContext)
 
-  describe('available requirement pagination and selection filtering', () => {
-    it('loads available requirements without sending the fixed status filter', async () => {
-      renderRequirementsSpecificationDetailClient()
+  registerAvailabilityTests(workflowContext)
 
-      await waitFor(() => {
-        expect(availableRequirementsFetchUrls().length).toBeGreaterThan(0)
-      })
+  registerMetadataTableTests(workflowContext)
 
-      const initialUrl = availableRequirementsFetchUrls()[0] ?? ''
-      const params = searchParamsFromPath(initialUrl)
-      expect(params.get('locale')).toBe('en')
-      expect(params.has('statuses')).toBe(false)
-    })
-
-    it('treats omitted optional list payload fields as an empty page', async () => {
-      availableRequirementsGetHandler = async () => okJson({})
-      specificationItemsGetHandler = async () => okJson({})
-      needsReferencesGetBody = {}
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        errors: [
-          {
-            key: SPECIFICATION_PRELOAD_ERROR_KEYS.needsReferences,
-            message: 'Needs references missing from preload',
-          },
-        ],
-      })
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('table', { name: 'available requirements' }),
-        ).toHaveTextContent('')
-      })
-      fireEvent.click(
-        screen.getByRole('button', { name: 'sort-description-items' }),
-      )
-      await waitFor(() => {
-        expect(screen.getByText('specification.noItems')).toBeInTheDocument()
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.newLocalRequirement',
-        }),
-      )
-      await screen.findByRole('dialog')
-      await waitFor(() => {
-        expect(
-          fetchMock.mock.calls.some(
-            ([input]) =>
-              (typeof input === 'string' ? input : input.url) ===
-              specificationApiPath('/needs-references'),
-          ),
-        ).toBe(true)
-      })
-    })
-
-    it('keeps requirement-selection filtering opt-in for available requirements', async () => {
-      availableRequirementsSelectionFilter = {
-        applied: false,
-        hasCurrentAnswers: true,
-        hasRequirementSelection: true,
-        hasNoRequirementSelection: false,
-        requirementIds: [202],
-      }
-      renderRequirementsSpecificationDetailClient()
-
-      const toggle = await screen.findByRole('switch', {
-        name: 'specification.filterWithRequirementSelectionQuestions',
-      })
-      expect(toggle).not.toBeChecked()
-      expect(
-        availableRequirementsFetchUrls().some(url =>
-          url.includes('applyRequirementSelectionFilter=true'),
-        ),
-      ).toBe(false)
-
-      fireEvent.click(toggle)
-
-      await waitFor(() => {
-        expect(
-          availableRequirementsFetchUrls().some(url =>
-            url.includes('applyRequirementSelectionFilter=true'),
-          ),
-        ).toBe(true)
-      })
-      expect(toggle).toBeChecked()
-    })
-
-    it('turns off requirement-selection filtering when refreshed answers no longer select requirements', async () => {
-      availableRequirementsSelectionFilter = {
-        applied: false,
-        hasCurrentAnswers: true,
-        hasRequirementSelection: true,
-        hasNoRequirementSelection: false,
-        requirementIds: [202],
-      }
-      availableRequirementsGetHandler = async url =>
-        okJson({
-          pagination: { hasMore: false },
-          requirements: [initialAvailableRequirement],
-          selectionFilter: url.includes('applyRequirementSelectionFilter=true')
-            ? {
-                applied: true,
-                hasCurrentAnswers: true,
-                hasRequirementSelection: false,
-                hasNoRequirementSelection: true,
-                requirementIds: [],
-              }
-            : availableRequirementsSelectionFilter,
-        })
-      renderRequirementsSpecificationDetailClient()
-      const toggle = await screen.findByRole('switch', {
-        name: 'specification.filterWithRequirementSelectionQuestions',
-      })
-      fireEvent.click(toggle)
-
-      await waitFor(() => {
-        expect(toggle).not.toBeChecked()
-        expect(toggle).toBeDisabled()
-      })
-      expect(
-        screen.getByText(
-          'specification.requirementSelectionNoPublishedMatches',
-        ),
-      ).toBeInTheDocument()
-    })
-
-    it('disables the requirement-selection filter toggle when answers provide no requirement selection', async () => {
-      availableRequirementsSelectionFilter = {
-        applied: false,
-        hasCurrentAnswers: true,
-        hasRequirementSelection: false,
-        hasNoRequirementSelection: true,
-        requirementIds: [],
-      }
-      renderRequirementsSpecificationDetailClient()
-
-      const toggle = await screen.findByRole('switch', {
-        name: 'specification.filterWithRequirementSelectionQuestions',
-      })
-      expect(toggle).toBeDisabled()
-      expect(toggle).not.toBeChecked()
-      expect(toggle).toHaveAttribute(
-        'title',
-        'specification.requirementSelectionFilterDisabledTooltip',
-      )
-    })
-
-    it('renders the requirement-selection toggle without a native input surface', async () => {
-      availableRequirementsSelectionFilter = {
-        applied: false,
-        hasCurrentAnswers: true,
-        hasRequirementSelection: true,
-        hasNoRequirementSelection: false,
-        requirementIds: [202],
-      }
-      renderRequirementsSpecificationDetailClient()
-
-      const toggle = await screen.findByRole('switch', {
-        name: 'specification.filterWithRequirementSelectionQuestions',
-      })
-      const switchTrack = toggle.querySelector('span[aria-hidden="true"]')
-      if (!(switchTrack instanceof HTMLElement)) {
-        throw new Error('Expected requirement-selection toggle track')
-      }
-
-      expect(toggle.tagName).toBe('BUTTON')
-      expect(toggle.className).not.toContain('focus-within:ring')
-      expect(toggle.className).not.toContain('absolute')
-      expect(toggle.className).not.toContain('inset-0')
-      expect(toggle.className).not.toContain('w-full')
-      expect(switchTrack.className).not.toContain('peer-focus-visible:ring')
-    })
-
-    it('keeps the requirement-selection toggle mounted while filtered requirements refresh', async () => {
-      availableRequirementsSelectionFilter = {
-        applied: false,
-        hasCurrentAnswers: true,
-        hasRequirementSelection: true,
-        hasNoRequirementSelection: false,
-        requirementIds: [202],
-      }
-      renderRequirementsSpecificationDetailClient()
-
-      const toggle = await screen.findByRole('switch', {
-        name: 'specification.filterWithRequirementSelectionQuestions',
-      })
-
-      let resolveFetch:
-        | ((value: { json: () => Promise<unknown>; ok: boolean }) => void)
-        | undefined
-      fetchMock.mockImplementationOnce(
-        () =>
-          new Promise(resolve => {
-            resolveFetch = resolve
-          }),
-      )
-
-      fireEvent.click(toggle)
-
-      await waitFor(() => {
-        expect(
-          availableRequirementsFetchUrls().some(url =>
-            url.includes('applyRequirementSelectionFilter=true'),
-          ),
-        ).toBe(true)
-      })
-
-      expect(
-        screen.getByRole('switch', {
-          name: 'specification.filterWithRequirementSelectionQuestions',
-        }),
-      ).toBeChecked()
-
-      await act(async () => {
-        resolveFetch?.(
-          okJson({
-            pagination: { hasMore: false },
-            requirements: [initialAvailableRequirement],
-            selectionFilter: {
-              applied: true,
-              hasCurrentAnswers: true,
-              hasRequirementSelection: true,
-              hasNoRequirementSelection: false,
-              requirementIds: [202],
-            },
-          }),
-        )
-      })
-    })
-
-    it('loads more available requirements without sending the fixed status filter', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        availableRequirements: {
-          hasMore: true,
-          nextCursor: 'cursor-1',
-          rows: [initialAvailableRequirement],
-        },
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'load-more-available' }),
-      )
-
-      await waitFor(() => {
-        expect(
-          availableRequirementsFetchUrls().some(
-            url => searchParamsFromPath(url).get('cursor') === 'cursor-1',
-          ),
-        ).toBe(true)
-      })
-
-      const params = searchParamsFromPath(
-        availableRequirementsFetchUrls().find(
-          url => searchParamsFromPath(url).get('cursor') === 'cursor-1',
-        ) ?? '',
-      )
-      expect(params.has('statuses')).toBe(false)
-    })
-
-    it('reloads available requirements and announces an invalid cursor', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        availableRequirements: {
-          hasMore: true,
-          nextCursor: 'cursor-1',
-          rows: [initialAvailableRequirement],
-        },
-      })
-      await waitForInitialAvailableRequirementsRefresh()
-      const requestCountBeforeLoadMore = availableRequirementsFetchUrls().length
-      fetchMock.mockImplementationOnce(() =>
-        Promise.resolve({
-          clone() {
-            return this
-          },
-          json: async () => ({ code: 'invalid_cursor' }),
-          ok: false,
-          status: 400,
-        } as Response),
-      )
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'load-more-available' }),
-      )
-
-      expect(
-        await screen.findByText('common.requirementListRefreshed'),
-      ).toHaveAttribute('role', 'status')
-      await waitFor(() => {
-        expect(availableRequirementsFetchUrls()).toHaveLength(
-          requestCountBeforeLoadMore + 2,
-        )
-      })
-      expect(
-        screen.getByRole('table', { name: 'available requirements' }),
-      ).toHaveTextContent('202')
-    })
-
-    it('ignores stale invalid-cursor recovery after available filters change', async () => {
-      availableRequirementsGetHandler = async url => {
-        const filtered = searchParamsFromPath(url).has('requirementPackageIds')
-        return okJson({
-          pagination: {
-            hasMore: !filtered,
-            nextCursor: filtered ? null : 'cursor-1',
-          },
-          requirements: [initialAvailableRequirement],
-        })
-      }
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        availableRequirements: {
-          hasMore: true,
-          nextCursor: 'cursor-1',
-          rows: [initialAvailableRequirement],
-        },
-        requirementPackages: [{ id: 1, name: 'Mobile use' }],
-      })
-      await waitForInitialAvailableRequirementsRefresh()
-      let resolveStaleLoadMore: ((response: Response) => void) | undefined
-      fetchMock.mockImplementationOnce(
-        () =>
-          new Promise<Response>(resolve => {
-            resolveStaleLoadMore = resolve
-          }),
-      )
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'load-more-available' }),
-      )
-      await waitFor(() => {
-        expect(
-          availableRequirementsFetchUrls().some(
-            url => searchParamsFromPath(url).get('cursor') === 'cursor-1',
-          ),
-        ).toBe(true)
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'filter-package-available-1' }),
-      )
-      await waitFor(() => {
-        expect(
-          availableRequirementsFetchUrls().some(url =>
-            url.includes('requirementPackageIds=1'),
-          ),
-        ).toBe(true)
-      })
-      const requestCountAfterFilter = availableRequirementsFetchUrls().length
-
-      await act(async () => {
-        resolveStaleLoadMore?.({
-          clone() {
-            return this
-          },
-          json: async () => ({ code: 'invalid_cursor' }),
-          ok: false,
-          status: 400,
-        } as Response)
-      })
-      expect(availableRequirementsFetchUrls()).toHaveLength(
-        requestCountAfterFilter,
-      )
-      expect(screen.queryByText('common.requirementListRefreshed')).toBeNull()
-    })
-  })
-
-  describe('metadata, layout, filters, and item status', () => {
-    it('opens and closes the specification edit dialog from the title action', async () => {
-      const { container } = renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      const headerSummary = container.querySelector(
-        '[data-specification-detail-header-summary="true"]',
-      )
-      const headerMetadata = container.querySelector(
-        '[data-specification-detail-header-metadata="true"]',
-      )
-      const pageShell = container.querySelector(
-        '[data-specification-detail-page-shell="true"]',
-      ) as HTMLDivElement | null
-      const splitPanel = container.querySelector(
-        '[data-specification-detail-split-panel="true"]',
-      ) as HTMLDivElement | null
-      const titleRow = container.querySelector(
-        '[data-specification-detail-title-row="true"]',
-      )
-      expect(headerSummary).toBeTruthy()
-      expect(headerMetadata).toBeTruthy()
-      expect(pageShell).toBeTruthy()
-      expect(splitPanel).toBeTruthy()
-      expect(titleRow).toBeTruthy()
-      expect(pageShell?.className).toContain('xl:h-[calc(100dvh-4rem)]')
-      expect(splitPanel?.className).toContain('xl:-mx-8')
-      expect(splitPanel?.className).toContain('xl:flex-1')
-      expect(
-        screen.queryByRole('link', { name: 'nav.specifications' }),
-      ).not.toBeInTheDocument()
-      expect(headerSummary).toHaveTextContent('Platform')
-      expect(headerSummary).toHaveTextContent('Ada Admin')
-      expect(headerSummary).toHaveTextContent('SE5560000001-ada1')
-      expect(headerSummary).toHaveTextContent('Program')
-      expect(headerSummary).toHaveTextContent('Shared IAM business case')
-      expect(headerSummary).not.toHaveTextContent(
-        'specification.businessNeedsReference',
-      )
-      expect(headerSummary).toHaveClass('xl:grid')
-      expect(headerSummary).toHaveClass(
-        'xl:grid-cols-[minmax(40vw,1fr)_minmax(0,1fr)]',
-      )
-      expect(headerMetadata).not.toHaveTextContent('Shared IAM business case')
-      expect(headerMetadata).toHaveClass('grid-flow-col')
-      expect(headerMetadata).toHaveClass('auto-cols-[minmax(12rem,1fr)]')
-      expect(headerMetadata).toHaveClass('overflow-x-auto')
-      expect(headerMetadata).toHaveClass('xl:auto-cols-fr')
-      expect(headerMetadata).not.toHaveClass('xl:grid-cols-3')
-
-      const editButton = screen.getByRole('button', {
-        name: /specification\.editSpecification/i,
-      })
-      expect(editButton).toHaveAttribute('aria-expanded', 'false')
-      expect(editButton).toHaveAttribute(
-        'data-developer-mode-name',
-        'detail action',
-      )
-      expect(editButton).toHaveAttribute(
-        'data-developer-mode-context',
-        'requirements specification detail',
-      )
-      expect(editButton).toHaveAttribute(
-        'data-developer-mode-value',
-        'edit specification',
-      )
-      expect(titleRow).toContainElement(
-        screen.getByRole('heading', {
-          level: 1,
-          name: 'Authorization and IAM',
-        }),
-      )
-      expect(titleRow).toContainElement(editButton)
-
-      fireEvent.click(editButton)
-
-      expect(editButton).toHaveAttribute('aria-expanded', 'true')
-      const dialog = screen.getByRole('dialog', {
-        name: /specification\.editSpecification/i,
-      })
-      expect(
-        within(dialog).getByRole('textbox', { name: /specification\.name/ }),
-      ).toHaveValue('Authorization and IAM')
-
-      const form = document.body.querySelector(
-        '[data-developer-mode-name="crud form"][data-developer-mode-context="requirements specification detail"]',
-      )
-      expect(form).toHaveAttribute('data-developer-mode-value', 'edit')
-
-      fireEvent.click(screen.getByRole('button', { name: /common\.cancel/i }))
-
-      await waitFor(() => {
-        expect(
-          screen.queryByRole('textbox', { name: /specification\.name/ }),
-        ).not.toBeInTheDocument()
-      })
-      expect(editButton).toHaveAttribute('aria-expanded', 'false')
-    })
-
-    it('saves specification metadata from the title edit action and refreshes it', async () => {
-      renderRequirementsSpecificationDetailClient()
-      const initialMetadataGetCount = fetchMock.mock.calls.filter(
-        ([url, init]) =>
-          url === specificationApiPath() &&
-          (init as RequestInit | undefined)?.method !== 'PUT',
-      ).length
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.editSpecification' }),
-      )
-      const dialog = await screen.findByRole('dialog', {
-        name: 'specification.editSpecification',
-      })
-      fireEvent.change(
-        within(dialog).getByRole('textbox', { name: /specification\.name/ }),
-        { target: { value: 'Updated specification name' } },
-      )
-      fireEvent.click(
-        within(dialog).getByRole('button', { name: /common\.save/i }),
-      )
-
-      await waitFor(() => {
-        const putCall = fetchMock.mock.calls.find(
-          ([url, init]) =>
-            url === specificationApiPath() &&
-            (init as RequestInit | undefined)?.method === 'PUT',
-        )
-        expect(
-          JSON.parse(String((putCall?.[1] as RequestInit | undefined)?.body)),
-        ).toEqual(
-          expect.objectContaining({ name: 'Updated specification name' }),
-        )
-      })
-      await waitFor(() =>
-        expect(
-          screen.queryByRole('dialog', {
-            name: 'specification.editSpecification',
-          }),
-        ).toBeNull(),
-      )
-      await waitFor(() => {
-        const metadataGetCount = fetchMock.mock.calls.filter(
-          ([url, init]) =>
-            url === specificationApiPath() &&
-            (init as RequestInit | undefined)?.method !== 'PUT',
-        ).length
-        expect(metadataGetCount).toBeGreaterThan(initialMetadataGetCount)
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Updated specification name',
-          }),
-        ).toBeInTheDocument()
-      })
-    })
-
-    it('moves to the not-found state when metadata disappears during an edit refresh', async () => {
-      specificationMetaReturnsNotFound = true
-      renderRequirementsSpecificationDetailClient()
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.editSpecification' }),
-      )
-      const dialog = await screen.findByRole('dialog', {
-        name: 'specification.editSpecification',
-      })
-      fireEvent.change(
-        within(dialog).getByRole('textbox', { name: /specification\.name/ }),
-        { target: { value: 'Specification awaiting removal' } },
-      )
-      fireEvent.click(
-        within(dialog).getByRole('button', { name: /common\.save/i }),
-      )
-
-      expect(
-        await screen.findByText('specification.specificationNotFound'),
-      ).toBeInTheDocument()
-      expect(
-        screen.getByRole('link', {
-          name: /specification\.backToSpecifications/,
-        }),
-      ).toHaveAttribute('href', '/specifications')
-    })
-
-    it('does not fail open when specification permissions are missing', async () => {
-      const { permissions: omittedPermissions, ...specWithoutPermissions } =
-        initialSpec
-      void omittedPermissions
-
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        spec: specWithoutPermissions,
-      })
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      expect(
-        screen.queryByRole('button', {
-          name: /specification\.editSpecification/i,
-        }),
-      ).not.toBeInTheDocument()
-    })
-
-    it('renders a minimal Swedish read-only specification without optional metadata', async () => {
-      intlState.locale = 'sv'
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        availableNeedsRefs: [
-          {
-            createdAt: '',
-            description: 'Beskrivning',
-            id: 81,
-            libraryItemCount: 0,
-            linkedItemCount: undefined,
-            specificationLocalRequirementCount: 0,
-            text: 'Behov',
-            updatedAt: '',
-          },
-        ],
-        spec: {
-          ...initialSpec,
-          businessNeedsReference: null,
-          governanceObjectType: null,
-          implementationType: null,
-          lifecycleStatus: null,
-          permissions: {
-            canEditContent: false,
-            canManageAssignments: false,
-            canReviewDecisions: false,
-            canUseAi: false,
-          },
-          responsibleDisplayName: 'Readonly Owner',
-          responsibleHsaId: '',
-        },
-      })
-
-      expect(
-        screen.queryByRole('button', {
-          name: 'specification.editSpecification',
-        }),
-      ).toBeNull()
-      expect(screen.queryByText('Shared IAM business case')).toBeNull()
-      expect(screen.getByText('Readonly Owner')).toBeInTheDocument()
-      expect(screen.queryByText('SE5560000001-ada1')).toBeNull()
-      fireEvent.click(
-        screen.getByRole('tab', { name: /specification\.needsReferences/ }),
-      )
-      expect(screen.getByText('Beskrivning')).toBeInTheDocument()
-      expect(
-        screen.queryByRole('button', {
-          name: 'specification.newNeedsReference',
-        }),
-      ).toBeNull()
-      await waitForInitialAvailableRequirementsRefresh()
-    })
-
-    it('shows the not-found state with a partial-data warning when specification metadata is absent', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        errors: [{ key: 'specification', message: 'Metadata unavailable' }],
-        spec: null,
-      })
-
-      expect(screen.getByRole('status')).toHaveTextContent(
-        'specification.partialDataLoadWarning',
-      )
-      expect(
-        screen.getByText('specification.specificationNotFound'),
-      ).toBeInTheDocument()
-      expect(
-        screen.getByRole('link', {
-          name: /specification\.backToSpecifications/,
-        }),
-      ).toHaveAttribute('href', '/specifications')
-      await waitForInitialAvailableRequirementsRefresh()
-    })
-
-    it('does not show a read-only notice for assignment-only managers', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        spec: {
-          ...initialSpec,
-          permissions: {
-            canEditContent: false,
-            canManageAssignments: true,
-            canReviewDecisions: false,
-            canUseAi: false,
-          },
-        },
-      })
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      expect(
-        screen.getByRole('button', {
-          name: /specification\.editSpecification/i,
-        }),
-      ).toBeInTheDocument()
-      expect(screen.queryByText('specification.readOnlyNotice')).toBeNull()
-    })
-
-    it('hides create actions but keeps output actions for read-only kravunderlag detail users', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        spec: {
-          ...initialSpec,
-          permissions: {
-            canEditContent: false,
-            canManageAssignments: false,
-            canReviewDecisions: false,
-            canUseAi: false,
-          },
-        },
-      })
-      await waitForInitialAvailableRequirementsRefresh()
-
-      const itemsTable = latestItemsTableProps()
-      expect(itemsTable.selectable).toBe(false)
-      expect(
-        screen.queryByRole('button', { name: 'select-row-101' }),
-      ).not.toBeInTheDocument()
-      const floatingActions = (itemsTable.floatingActions ?? []) as Array<{
-        hidden?: boolean
-        id: string
-        menuItems?: Array<{ id: string; kind?: string }>
-      }>
-
-      expect(floatingActions.map(action => action.id)).toEqual(['more-actions'])
-      expect(floatingActions[0]?.hidden).toBe(false)
-      expect(floatingActions[0]?.menuItems?.map(item => item.id)).toEqual([
-        'pdf-progress',
-        'pdf-traceability',
-        'separator-export-actions',
-        'export-full',
-      ])
-      expect(
-        floatingActions[0]?.menuItems?.some(
-          item => item.id === 'ai-assist-local',
-        ),
-      ).toBe(false)
-      expect(
-        floatingActions[0]?.menuItems?.some(item => item.id === 'import-local'),
-      ).toBe(false)
-    })
-
-    it('shows direct create and more actions but no columns action in the editable empty state', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        specificationItems: createSpecificationItemsPage([]),
-      })
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('button', {
-            name: 'specification.newLocalRequirement',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      expect(
-        screen.getByRole('button', { name: 'common.moreActions' }),
-      ).toBeInTheDocument()
-      expect(
-        screen.queryByRole('button', { name: 'common.columns' }),
-      ).not.toBeInTheDocument()
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'common.moreActions' }),
-      )
-
-      expect(
-        await screen.findByRole('menuitem', {
-          name: 'specification.aiGenerate',
-        }),
-      ).toBeInTheDocument()
-      expect(
-        screen.getByRole('menuitem', {
-          name: 'specification.importLocalRequirements',
-        }),
-      ).toBeInTheDocument()
-      expect(screen.queryByRole('separator')).not.toBeInTheDocument()
-    })
-
-    it('shows no empty-state actions to a read-only user', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        spec: {
-          ...initialSpec,
-          permissions: {
-            canEditContent: false,
-            canManageAssignments: false,
-            canReviewDecisions: false,
-            canUseAi: false,
-          },
-        },
-        specificationItems: createSpecificationItemsPage([]),
-      })
-
-      expect(screen.getByText('specification.noItems')).toBeInTheDocument()
-      expect(
-        screen.queryByRole('button', {
-          name: 'specification.newLocalRequirement',
-        }),
-      ).toBeNull()
-      expect(
-        screen.queryByRole('button', { name: 'common.moreActions' }),
-      ).toBeNull()
-      await waitForInitialAvailableRequirementsRefresh()
-    })
-
-    it('ignores stale and invalid stored detail column ids', async () => {
-      window.localStorage.setItem(
-        'requirements-specifications.visibleColumns.left.v2',
-        JSON.stringify(['uniqueId']),
-      )
-      window.localStorage.setItem(
-        'requirements-specifications.visibleColumns.left.v3',
-        JSON.stringify(['uniqueId', 'unknownSpecificationColumn']),
-      )
-
-      renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(requirementsTableMock).toHaveBeenCalled()
-      })
-
-      const leftTableProps = requirementsTableMock.mock.calls.find(
-        ([props]) => props.stickyTitle,
-      )?.[0]
-
-      expect(leftTableProps?.visibleColumns).toEqual([
-        'uniqueId',
-        'description',
-        'area',
-        'needsReference',
-      ])
-    })
-
-    it('passes context-specific reset defaults to the detail tables', async () => {
-      renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(requirementsTableMock.mock.calls.length).toBeGreaterThanOrEqual(
-          2,
-        )
-      })
-
-      const tableProps = requirementsTableMock.mock.calls.map(call => call[0])
-      const leftTableProps = tableProps.find(
-        props => props.rows[0]?.id === initialSpecificationItem.id,
-      )
-      const rightTableProps = tableProps.find(
-        props => props.rows[0]?.id === initialAvailableRequirement.id,
-      )
-
-      expect(leftTableProps?.defaultVisibleColumns).toEqual([
-        'uniqueId',
-        'description',
-        'area',
-        'needsReference',
-      ])
-      expect(rightTableProps?.defaultVisibleColumns).toEqual([
-        'uniqueId',
-        'description',
-        'area',
-      ])
-    })
-
-    it('expands both tables, refreshes details, updates status, and persists columns', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        specificationItemStatuses: [
-          {
-            color: '#22c55e',
-            descriptionEn: null,
-            descriptionSv: null,
-            iconName: null,
-            id: 2,
-            nameEn: 'Included',
-            nameSv: 'Inkluderad',
-            sortOrder: 2,
-          },
-        ],
-      })
-      await waitForInitialAvailableRequirementsRefresh()
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'expand-row-items-101' }),
-      )
-      expect(
-        await screen.findByText('Requirement detail 101'),
-      ).toBeInTheDocument()
-      const itemsFetchCountBeforeRefresh = fetchMock.mock.calls.filter(
-        ([input]) => String(input).startsWith(specificationApiPath('/items?')),
-      ).length
-      fireEvent.click(
-        screen.getByRole('button', { name: 'refresh requirement detail' }),
-      )
-      await waitFor(() => {
-        expect(
-          fetchMock.mock.calls.filter(([input]) =>
-            String(input).startsWith(specificationApiPath('/items?')),
-          ).length,
-        ).toBeGreaterThan(itemsFetchCountBeforeRefresh)
-      })
-      fireEvent.click(screen.getByRole('button', { name: 'set-status-lib:31' }))
-      fireEvent.click(screen.getByRole('button', { name: 'set-columns-items' }))
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'expand-row-available-202' }),
-      )
-      expect(
-        await screen.findByText('Requirement detail 202'),
-      ).toBeInTheDocument()
-      fireEvent.click(
-        screen.getByRole('button', { name: 'set-columns-available' }),
-      )
-      fireEvent.click(
-        screen.getByRole('button', { name: 'expand-row-available-202' }),
-      )
-      expect(
-        screen.queryByText('Requirement detail 202'),
-      ).not.toBeInTheDocument()
-
-      await waitFor(() => {
-        expect(fetchMock).toHaveBeenCalledWith(
-          '/api/requirements-specifications/8/items/lib%3A31',
-          expect.objectContaining({ method: 'PATCH' }),
-        )
-      })
-      await waitFor(() => {
-        expect(
-          window.localStorage.getItem(
-            'requirement-specifications.visibleColumns.left.v1',
-          ),
-        ).toBe('["uniqueId"]')
-        expect(
-          window.localStorage.getItem(
-            'requirement-specifications.visibleColumns.right.v1',
-          ),
-        ).toBe('["uniqueId"]')
-      })
-    })
-
-    it('restores the original item when a usage-status update fails', async () => {
-      itemStatusPatchOk = false
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        specificationItemStatuses: [
-          {
-            color: '#22c55e',
-            descriptionEn: null,
-            descriptionSv: null,
-            iconName: null,
-            id: 2,
-            nameEn: 'Included',
-            nameSv: 'Inkluderad',
-            sortOrder: 2,
-          },
-        ],
-      })
-
-      fireEvent.click(screen.getByRole('button', { name: 'set-status-lib:31' }))
-
-      await waitFor(() => {
-        expect(latestItemsTableProps().rows[0]).toEqual(
-          initialSpecificationItem,
-        )
-      })
-    })
-
-    it('ignores a usage-status choice that is not in the specification catalog', async () => {
-      renderRequirementsSpecificationDetailClient()
-      fireEvent.click(screen.getByRole('button', { name: 'set-status-lib:31' }))
-
-      expect(
-        fetchMock.mock.calls.some(
-          ([url, init]) =>
-            String(url).includes('/items/lib%3A31') &&
-            (init as RequestInit | undefined)?.method === 'PATCH',
-        ),
-      ).toBe(false)
-      expect(
-        screen.getByRole('table', { name: 'items requirements' }),
-      ).toHaveTextContent('lib:31')
-      await waitForInitialAvailableRequirementsRefresh()
-    })
-
-    it('submits an explicit needs-reference clear even for an unknown item', async () => {
-      renderRequirementsSpecificationDetailClient()
-      fireEvent.click(
-        screen.getByRole('button', { name: 'clear-needs-ref-lib:missing' }),
-      )
-
-      await waitFor(() => {
-        expect(fetchMock).toHaveBeenCalledWith(
-          `/api/requirements-specifications/8/items/${encodeURIComponent('lib:missing')}`,
-          expect.objectContaining({
-            body: JSON.stringify({ needsReferenceId: null }),
-            method: 'PATCH',
-          }),
-        )
-      })
-      expect(
-        screen.getByRole('table', { name: 'items requirements' }),
-      ).toHaveTextContent('lib:31')
-    })
-
-    it('expands and refreshes a specification-local requirement detail', async () => {
-      const localItem = {
-        ...initialSpecificationItem,
-        id: 401,
-        isSpecificationLocal: true,
-        itemRef: 'local:401',
-        kind: 'specificationLocal' as const,
-        specificationItemId: undefined,
-        specificationLocalRequirementId: 401,
-        uniqueId: 'KRAV0001',
-      }
-      specificationItemsGetItems = [localItem]
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        specificationItems: createSpecificationItemsPage([localItem]),
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'expand-row-items-401' }),
-      )
-      const localDetail = await screen.findByRole('button', {
-        name: 'Local requirement detail 401',
-      })
-      const itemsFetchCountBeforeRefresh = fetchMock.mock.calls.filter(
-        ([input]) => String(input).startsWith(specificationApiPath('/items?')),
-      ).length
-      fireEvent.click(localDetail)
-
-      await waitFor(() => {
-        expect(
-          fetchMock.mock.calls.filter(([input]) =>
-            String(input).startsWith(specificationApiPath('/items?')),
-          ).length,
-        ).toBeGreaterThan(itemsFetchCountBeforeRefresh)
-      })
-      expect(localRequirementDetailState.renderSpy).toHaveBeenCalledWith(
-        expect.objectContaining({ localRequirementId: 401 }),
-      )
-    })
-
-    it('shows generic requirement detail when an application has no persisted membership id', async () => {
-      const unresolvedItem = {
-        ...initialSpecificationItem,
-        itemRef: 'external:101',
-        specificationItemId: undefined,
-      }
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        specificationItems: createSpecificationItemsPage([unresolvedItem]),
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'expand-row-items-101' }),
-      )
-      expect(screen.getByText('Requirement detail 101')).toBeInTheDocument()
-      expect(requirementDetailState.renderSpy).toHaveBeenLastCalledWith(
-        expect.objectContaining({ requirementId: 101 }),
-      )
-      fireEvent.click(
-        screen.getByRole('button', { name: 'expand-row-items-101' }),
-      )
-      expect(
-        screen.queryByText('Requirement detail 101'),
-      ).not.toBeInTheDocument()
-      await waitForInitialAvailableRequirementsRefresh()
-    })
-
-    it('loads persisted detail columns after the hydration-safe default render', async () => {
-      const storedLeftColumns = [
-        'uniqueId',
-        'description',
-        'area',
-        'needsReference',
-        'status',
-      ]
-      const storedRightColumns = [
-        'uniqueId',
-        'description',
-        'area',
-        'status',
-        'type',
-      ]
-      window.localStorage.setItem(
-        'requirement-specifications.visibleColumns.left.v1',
-        JSON.stringify(storedLeftColumns),
-      )
-      window.localStorage.setItem(
-        'requirement-specifications.visibleColumns.right.v1',
-        JSON.stringify(storedRightColumns),
-      )
-
-      renderRequirementsSpecificationDetailClient()
-
-      const firstLeftTableProps = requirementsTableMock.mock.calls.find(
-        ([props]) => props.rows[0]?.id === initialSpecificationItem.id,
-      )?.[0]
-
-      expect(firstLeftTableProps?.visibleColumns).toEqual([
-        'uniqueId',
-        'description',
-        'area',
-        'needsReference',
-      ])
-
-      await waitFor(() => {
-        const latestCalls = [...requirementsTableMock.mock.calls].reverse()
-        const latestLeftTableProps = latestCalls.find(
-          ([props]) => props.rows[0]?.id === initialSpecificationItem.id,
-        )?.[0]
-        const latestRightTableProps = latestCalls.find(
-          ([props]) => props.rows[0]?.id === initialAvailableRequirement.id,
-        )?.[0]
-
-        expect(latestLeftTableProps?.visibleColumns).toEqual(storedLeftColumns)
-        expect(latestRightTableProps?.visibleColumns).toEqual(
-          storedRightColumns,
-        )
-      })
-      expect(
-        window.localStorage.getItem(
-          'requirement-specifications.visibleColumns.left.v1',
-        ),
-      ).toBe(JSON.stringify(storedLeftColumns))
-    })
-
-    it('uses inline top rails and embeds the split panel tabs in sticky headers', async () => {
-      const { container } = renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(requirementsTableMock.mock.calls.length).toBeGreaterThanOrEqual(
-          2,
-        )
-      })
-
-      expect(
-        screen.queryByText('specification.itemsInSpecification', {
-          selector: 'h2',
-        }),
-      ).not.toBeInTheDocument()
-      expect(
-        screen.queryByText('specification.availableRequirements', {
-          selector: 'h2',
-        }),
-      ).not.toBeInTheDocument()
-      const stickyTitles = screen.getAllByTestId(
-        'requirements-table-sticky-title',
-      )
-      const leftStickyTitle = stickyTitles.find(element =>
-        within(element).queryByRole('tab', {
-          name: /specification\.itemsInSpecification/,
-        }),
-      )
-      const rightStickyTitle = stickyTitles.find(element =>
-        within(element).queryByRole('tab', {
-          name: /specification\.availableRequirements/,
-        }),
-      )
-
-      expect(leftStickyTitle).toBeTruthy()
-      expect(rightStickyTitle).toBeTruthy()
-      expect(
-        within(leftStickyTitle as HTMLElement).getByRole('tablist', {
-          name: 'specification.leftPanelTabs',
-        }),
-      ).toBeInTheDocument()
-      expect(
-        within(leftStickyTitle as HTMLElement).getByRole('tab', {
-          name: /specification\.needsReferences/,
-        }),
-      ).toBeInTheDocument()
-      expect(
-        within(rightStickyTitle as HTMLElement).getByRole('tablist', {
-          name: 'specification.rightPanelTabs',
-        }),
-      ).toBeInTheDocument()
-      expect(
-        within(rightStickyTitle as HTMLElement).getByRole('tab', {
-          name: /specification\.availableRequirements/,
-        }),
-      ).toHaveAttribute('aria-controls', 'right-panel-available')
-      const questionsTab = within(rightStickyTitle as HTMLElement).getByRole(
-        'tab',
-        {
-          name: /specification\.requirementSelectionQuestions/,
-        },
-      )
-      expect(questionsTab).toHaveAttribute(
-        'aria-controls',
-        'right-panel-questions',
-      )
-
-      fireEvent.click(questionsTab)
-
-      await waitFor(() => {
-        expect(
-          screen.getByText('specificationRequirementSelection.noQuestions'),
-        ).toBeInTheDocument()
-      })
-      const questionsPanel = container.querySelector('#right-panel-questions')
-
-      expect(questionsPanel).toBeTruthy()
-      expect(
-        within(questionsPanel as HTMLElement).getByRole('tablist', {
-          name: 'specification.rightPanelTabs',
-        }),
-      ).toBeInTheDocument()
-      expect(
-        within(questionsPanel as HTMLElement).queryByText(
-          'specificationRequirementSelection.title',
-          { selector: 'h2' },
-        ),
-      ).not.toBeInTheDocument()
-
-      const tableProps = requirementsTableMock.mock.calls.map(call => call[0])
-
-      expect(
-        tableProps.every(
-          props => props.floatingActionRailPlacement === 'inline-top',
-        ),
-      ).toBe(true)
-      expect(
-        tableProps.every(props => props.stickyTopOffsetClassName === 'top-0'),
-      ).toBe(true)
-      expect(
-        container.querySelector(
-          '[data-specification-detail-list-panel="items"]',
-        ),
-      ).toBeTruthy()
-      expect(
-        container.querySelector(
-          '[data-specification-detail-list-panel="available"]',
-        ),
-      ).toBeTruthy()
-    })
-
-    it('filters requirement applications when a requirement package chip is selected', async () => {
-      const requirementPackages = [
-        { id: 1, name: 'Mobile use' },
-        { id: 2, name: 'Operations' },
-      ]
-      const firstItem = {
-        ...initialSpecificationItem,
-        requirementPackageIds: [1],
-      }
-      const secondItem = {
-        ...initialSpecificationItem,
-        id: 102,
-        itemRef: 'lib:32',
-        requirementPackageIds: [2],
-        specificationItemId: 32,
-        uniqueId: 'BEH0002',
-        version: {
-          ...initialSpecificationItem.version,
-          description: 'Operational monitoring should be in place.',
-        },
-      }
-      specificationItemsGetItems = [firstItem, secondItem]
-
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        leftRequirementPackageCatalog:
-          createRequirementPackageCatalogPage(requirementPackages),
-        requirementPackages,
-        specificationItems: createSpecificationItemsPage([
-          firstItem,
-          secondItem,
-        ]),
-      })
-
-      await waitFor(() => {
-        expect(requirementRowNames('items')).toEqual(['lib:31', 'lib:32'])
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'filter-package-items-1' }),
-      )
-
-      await waitFor(() => {
-        const latestItemsProps = [...requirementsTableMock.mock.calls]
-          .reverse()
-          .find(
-            ([props]) => !props.excludeColumns?.includes('needsReference'),
-          )?.[0] as
-          | {
-              filterValues?: { requirementPackageIds?: number[] }
-              rows: { id: number }[]
-            }
-          | undefined
-
-        expect(latestItemsProps?.filterValues).toEqual({
-          requirementPackageIds: [1],
-        })
-        expect(latestItemsProps?.rows.map(row => row.id)).toEqual([101])
-      })
-      expect(
-        screen.getByRole('table', { name: 'items requirements' }),
-      ).toHaveTextContent('lib:31')
-      expect(
-        screen.getByRole('table', { name: 'items requirements' }),
-      ).not.toHaveTextContent('lib:32')
-    })
-
-    it('refreshes left norm-reference options when usage-status filters change', async () => {
-      const requestedStatuses: string[][] = []
-      normReferencesGetHandler = async url => {
-        requestedStatuses.push(searchParamsFromPath(url).getAll('statuses'))
-        return okJson({})
-      }
-      renderRequirementsSpecificationDetailClient()
-      fireEvent.click(
-        screen.getByRole('button', { name: 'toggle-status-filter-items' }),
-      )
-
-      await waitFor(() => {
-        expect(latestItemsTableProps().normReferences).toEqual([])
-        expect(
-          fetchMock.mock.calls.some(([input]) =>
-            String(typeof input === 'string' ? input : input.url).includes(
-              'statuses=3',
-            ),
-          ),
-        ).toBe(true)
-      })
-      fireEvent.click(
-        screen.getByRole('button', { name: 'toggle-status-filter-items' }),
-      )
-      await waitFor(() => {
-        expect(requestedStatuses).toContainEqual([])
-      })
-    })
-
-    it('keeps the item list usable when filtered norm-reference options fail', async () => {
-      normReferencesGetHandler = async () => {
-        throw 'Norm options unavailable'
-      }
-      renderRequirementsSpecificationDetailClient()
-      fireEvent.click(
-        screen.getByRole('button', { name: 'toggle-status-filter-items' }),
-      )
-
-      expect(await screen.findByRole('status')).toHaveTextContent(
-        'specification.loadNormReferencesFailed',
-      )
-      expect(
-        screen.getByRole('table', { name: 'items requirements' }),
-      ).toHaveTextContent('lib:31')
-    })
-
-    it('uses independent compact package filters with distinct server catalogs', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        leftRequirementPackageCatalog: createRequirementPackageCatalogPage([
-          { id: 2, name: 'Specification package' },
-        ]),
-        requirementPackages: [
-          { id: 1, name: 'Library package' },
-          { id: 2, name: 'Specification package' },
-        ],
-      })
-      await waitForInitialAvailableRequirementsRefresh()
-
-      expect(latestItemsTableProps()).toEqual(
-        expect.objectContaining({
-          requirementPackageCatalogStatus: 'loaded',
-          requirementPackageFilterPresentation: 'compact-band',
-          requirementPackages: [{ id: 2, name: 'Specification package' }],
-        }),
-      )
-      expect(latestAvailableTableProps()).toEqual(
-        expect.objectContaining({
-          requirementPackageCatalogStatus: 'loaded',
-          requirementPackageFilterPresentation: 'compact-band',
-          requirementPackages: [
-            { id: 1, name: 'Library package' },
-            { id: 2, name: 'Specification package' },
-          ],
-        }),
-      )
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'filter-package-items-2' }),
-      )
-      fireEvent.click(
-        screen.getByRole('button', { name: 'filter-package-available-1' }),
-      )
-
-      await waitFor(() => {
-        expect(latestItemsTableProps().filterValues).toEqual({
-          requirementPackageIds: [2],
-        })
-        expect(latestAvailableTableProps().filterValues).toEqual({
-          requirementPackageIds: [1],
-        })
-      })
-    })
-
-    it('keeps the item list usable while traversing every catalog page independently', async () => {
-      let resolveSecondPage: ((value: unknown) => void) | undefined
-      specificationRequirementPackagesGetHandler = async url => {
-        expect(searchParamsFromPath(url).get('cursor')).toBe('package-page-2')
-        return new Promise(resolve => {
-          resolveSecondPage = resolve
-        })
-      }
-
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        leftRequirementPackageCatalog: createRequirementPackageCatalogPage(
-          [{ id: 1, name: 'First package' }],
-          { hasMore: true, nextCursor: 'package-page-2' },
-        ),
-      })
-
-      expect(
-        screen.getByRole('table', { name: 'items requirements' }),
-      ).toHaveTextContent('lib:31')
-      await waitFor(() => {
-        expect(latestItemsTableProps()).toEqual(
-          expect.objectContaining({
-            requirementPackageCatalogStatus: 'loading',
-            requirementPackages: [{ id: 1, name: 'First package' }],
-          }),
-        )
-      })
-
-      await act(async () => {
-        resolveSecondPage?.(
-          okJson({
-            pagination: {
-              count: 1,
-              hasMore: false,
-              limit: 50,
-              nextCursor: null,
-            },
-            requirementPackages: [{ id: 51, name: 'Later package' }],
-            selectedRequirementPackages: [],
-          }),
-        )
-      })
-
-      await waitFor(() => {
-        expect(latestItemsTableProps()).toEqual(
-          expect.objectContaining({
-            requirementPackageCatalogStatus: 'loaded',
-            requirementPackages: [
-              { id: 1, name: 'First package' },
-              { id: 51, name: 'Later package' },
-            ],
-          }),
-        )
-      })
-    })
-
-    it('distinguishes failed package catalogs from successful empty catalogs', async () => {
-      const initialData = createInitialData()
-      initialData.errors = [
-        {
-          key: SPECIFICATION_PRELOAD_ERROR_KEYS.requirementPackages,
-          message: 'right failed',
-        },
-        {
-          key: SPECIFICATION_PRELOAD_ERROR_KEYS.specificationRequirementPackages,
-          message: 'left failed',
-        },
-      ]
-
-      renderRequirementsSpecificationDetailClient(initialData)
-      await waitForInitialAvailableRequirementsRefresh()
-
-      expect(latestItemsTableProps()).toEqual(
-        expect.objectContaining({
-          requirementPackageCatalogStatus: 'failed',
-          requirementPackages: [],
-        }),
-      )
-      expect(latestAvailableTableProps()).toEqual(
-        expect.objectContaining({
-          requirementPackageCatalogStatus: 'failed',
-          requirementPackages: [],
-        }),
-      )
-    })
-
-    it('fails a malformed preloaded package continuation without issuing an ambiguous request', async () => {
-      const initialData = createInitialData()
-      initialData.leftRequirementPackageCatalog =
-        createRequirementPackageCatalogPage([], {
-          hasMore: true,
-          nextCursor: null,
-        })
-      renderRequirementsSpecificationDetailClient(initialData)
-
-      await waitFor(() => {
-        expect(latestItemsTableProps().requirementPackageCatalogStatus).toBe(
-          'failed',
-        )
-        expect(
-          screen.getByText('specification.loadRequirementPackagesFailed'),
-        ).toBeInTheDocument()
-      })
-      expect(
-        fetchMock.mock.calls.some(([input]) =>
-          (typeof input === 'string' ? input : input.url).startsWith(
-            specificationApiPath('/requirement-packages'),
-          ),
-        ),
-      ).toBe(false)
-    })
-
-    it('fails a package catalog page whose continuation cursor is missing', async () => {
-      const initialData = createInitialData()
-      initialData.leftRequirementPackageCatalog =
-        createRequirementPackageCatalogPage([], {
-          hasMore: true,
-          nextCursor: 'catalog-page-2',
-        })
-      specificationRequirementPackagesGetHandler = async () =>
-        okJson({
-          pagination: { hasMore: true, nextCursor: null },
-          requirementPackages: [],
-          selectedRequirementPackages: [],
-        })
-      renderRequirementsSpecificationDetailClient(initialData)
-
-      await waitFor(() => {
-        expect(latestItemsTableProps().requirementPackageCatalogStatus).toBe(
-          'failed',
-        )
-      })
-      expect(
-        screen.getByText('specification.loadRequirementPackagesFailed'),
-      ).toBeInTheDocument()
-    })
-
-    it('fails a package catalog whose continuation cursor repeats', async () => {
-      const initialData = createInitialData()
-      initialData.leftRequirementPackageCatalog =
-        createRequirementPackageCatalogPage([], {
-          hasMore: true,
-          nextCursor: 'catalog-repeat',
-        })
-      let requestCount = 0
-      specificationRequirementPackagesGetHandler = async () => {
-        requestCount += 1
-        return okJson({
-          pagination: { hasMore: true, nextCursor: 'catalog-repeat' },
-          requirementPackages: [],
-          selectedRequirementPackages: [],
-        })
-      }
-      renderRequirementsSpecificationDetailClient(initialData)
-
-      await waitFor(() => {
-        expect(latestItemsTableProps().requirementPackageCatalogStatus).toBe(
-          'failed',
-        )
-      })
-      expect(requestCount).toBe(2)
-    })
-
-    it('prunes a resolved package while preserving a selection made during refresh', async () => {
-      const packageOption = { id: 9, name: 'Current package' }
-      const replacementPackageOption = { id: 10, name: 'Replacement package' }
-      const item = {
-        ...initialSpecificationItem,
-        requirementPackageIds: [9],
-      }
-      const remainingItem = {
-        ...initialSpecificationItem,
-        id: 102,
-        itemRef: 'library:32',
-        requirementPackageIds: [10],
-        uniqueId: 'REQ-002',
-      }
-      specificationItemsGetItems = [item, remainingItem]
-      let resolveCatalogRefresh: (() => void) | undefined
-      specificationItemsGetHandler = async url => {
-        const packageIds = searchParamsFromPath(url)
-          .getAll('requirementPackageIds')
-          .map(Number)
-        const items =
-          packageIds.length > 0
-            ? specificationItemsGetItems.filter(candidate =>
-                candidate.requirementPackageIds?.some(id =>
-                  packageIds.includes(id),
-                ),
-              )
-            : specificationItemsGetItems
-        return okJson({
-          items,
-          pagination: {
-            count: items.length,
-            hasMore: false,
-            limit: 50,
-            nextCursor: null,
-          },
-        })
-      }
-      specificationRequirementPackagesGetHandler = async url =>
-        new Promise(resolve => {
-          expect(searchParamsFromPath(url).getAll('includeIds')).toEqual(['9'])
-          const response = okJson({
-            pagination: {
-              count: 0,
-              hasMore: false,
-              limit: 50,
-              nextCursor: null,
-            },
-            requirementPackages: [],
-            selectedRequirementPackages: [],
-          })
-          resolveCatalogRefresh = () => resolve(response)
-        })
-
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        leftRequirementPackageCatalog: createRequirementPackageCatalogPage([
-          packageOption,
-          replacementPackageOption,
-        ]),
-        requirementPackages: [packageOption],
-        specificationItems: createSpecificationItemsPage([item, remainingItem]),
-      })
-      await waitForInitialAvailableRequirementsRefresh()
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'filter-package-items-9' }),
-      )
-      await waitFor(() => {
-        expect(latestItemsTableProps().filterValues).toEqual({
-          requirementPackageIds: [9],
-        })
-      })
-
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-101' }))
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.removeSelected' }),
-      )
-      fireEvent.click(
-        await screen.findByRole('button', { name: 'common.delete' }),
-      )
-
-      await waitFor(() => {
-        expect(fetchMock).toHaveBeenCalledWith(
-          '/api/requirements-specifications/8/items',
-          expect.objectContaining({ method: 'DELETE' }),
-        )
-      })
-      await waitFor(() => {
-        expect(latestItemsTableProps()).toEqual(
-          expect.objectContaining({
-            requirementPackageCatalogStatus: 'loading',
-            requirementPackages: [packageOption, replacementPackageOption],
-          }),
-        )
-      })
-      fireEvent.click(
-        screen.getByRole('button', { name: 'filter-package-items-10' }),
-      )
-      await waitFor(() => {
-        expect(latestItemsTableProps().filterValues).toEqual({
-          requirementPackageIds: [9, 10],
-        })
-      })
-      act(() => resolveCatalogRefresh?.())
-      await waitFor(() => {
-        expect(latestItemsTableProps().requirementPackages).toEqual([])
-        expect(
-          latestItemsTableProps().filterValues?.requirementPackageIds,
-        ).toEqual([10])
-        expect(latestItemsTableProps().rows).toEqual([remainingItem])
-      })
-      expect(latestAvailableTableProps().requirementPackages).toEqual([
-        packageOption,
-      ])
-    })
-
-    it('marks a failed package catalog refresh as failed without treating stale options as loaded', async () => {
-      const packageOption = { id: 9, name: 'Stale package' }
-      const item = {
-        ...initialSpecificationItem,
-        requirementPackageIds: [9],
-      }
-      const remainingItem = {
-        ...initialSpecificationItem,
-        id: 102,
-        itemRef: 'lib:32',
-        requirementPackageIds: [],
-        specificationItemId: 32,
-        uniqueId: 'BEH0002',
-      }
-      specificationItemsGetItems = [item, remainingItem]
-      specificationItemsGetHandler = async () =>
-        okJson({
-          items: specificationItemsGetItems,
-          pagination: {
-            count: specificationItemsGetItems.length,
-            hasMore: false,
-            limit: 50,
-            nextCursor: null,
-          },
-        })
-      specificationRequirementPackagesGetHandler = async () => ({
-        json: async () => ({ error: 'facet unavailable' }),
-        ok: false,
-      })
-
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        leftRequirementPackageCatalog: createRequirementPackageCatalogPage([
-          packageOption,
-        ]),
-        requirementPackages: [packageOption],
-        specificationItems: createSpecificationItemsPage([item, remainingItem]),
-      })
-      await waitForInitialAvailableRequirementsRefresh()
-
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-101' }))
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.removeSelected' }),
-      )
-      fireEvent.click(
-        await screen.findByRole('button', { name: 'common.delete' }),
-      )
-
-      await waitFor(() => {
-        expect(latestItemsTableProps()).toEqual(
-          expect.objectContaining({
-            requirementPackageCatalogStatus: 'failed',
-            requirementPackages: [],
-          }),
-        )
-      })
-    })
-
-    it('sorts the complete requirement application list in both directions', async () => {
-      const firstItem = {
-        ...initialSpecificationItem,
-        version: {
-          ...initialSpecificationItem.version,
-          description: 'Zulu requirement',
-        },
-      }
-      const secondItem = {
-        ...initialSpecificationItem,
-        id: 102,
-        itemRef: 'lib:32',
-        specificationItemId: 32,
-        uniqueId: 'BEH0002',
-        version: {
-          ...initialSpecificationItem.version,
-          description: 'Alpha requirement',
-        },
-      }
-      specificationItemsGetItems = [firstItem, secondItem]
-
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        specificationItems: createSpecificationItemsPage([
-          firstItem,
-          secondItem,
-        ]),
-      })
-
-      await waitFor(() => {
-        expect(requirementRowNames('items')).toEqual(['lib:31', 'lib:32'])
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'sort-description-items' }),
-      )
-      await waitFor(() => {
-        expect(requirementRowNames('items')).toEqual(['lib:32', 'lib:31'])
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'sort-description-items' }),
-      )
-      await waitFor(() => {
-        expect(requirementRowNames('items')).toEqual(['lib:31', 'lib:32'])
-      })
-    })
-  })
-
-  describe('adding, importing, and creating requirement applications', () => {
-    it('keeps the add dialog open and shows inline errors when adding requirements fails', async () => {
-      addRequirementsResponse = {
-        body: { error: 'Could not add requirements' },
-        ok: false,
-      }
-
-      renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-
-      const dialog = await screen.findByRole('dialog')
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.confirmAdd' }),
-      )
-
-      expect(await screen.findByRole('alert')).toHaveTextContent(
-        'Could not add requirements',
-      )
-      expect(dialog).toBeInTheDocument()
-    })
-
-    it('adds selected requirements with an existing needs reference', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        availableNeedsRefs: [
-          {
-            createdAt: '',
-            description: 'Existing context',
-            id: 81,
-            libraryItemCount: 0,
-            linkedItemCount: 0,
-            specificationLocalRequirementCount: 0,
-            text: 'IAM-42',
-            updatedAt: '',
-          },
-        ],
-      })
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-      fireEvent.change(screen.getByLabelText('specification.addNeedsRef'), {
-        target: { value: '81' },
-      })
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.confirmAdd' }),
-      )
-
-      await waitFor(() => {
-        const postCall = fetchMock.mock.calls.find(
-          ([url, init]) =>
-            url === specificationApiPath('/items') &&
-            (init as RequestInit | undefined)?.method === 'POST',
-        )
-        expect(
-          JSON.parse(String((postCall?.[1] as RequestInit | undefined)?.body)),
-        ).toEqual({ needsReferenceId: 81, requirementIds: [202] })
-      })
-    })
-
-    it('adds selected requirements without a needs reference', async () => {
-      renderRequirementsSpecificationDetailClient()
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-      const help = await screen.findByRole('button', {
-        name: 'common.help: specification.addNeedsRef',
-      })
-      fireEvent.click(help)
-      expect(
-        screen.getByText('specification.addNeedsRefHelp'),
-      ).toBeInTheDocument()
-      fireEvent.click(help)
-      expect(screen.queryByText('specification.addNeedsRefHelp')).toBeNull()
-      const needsReferenceSelect = screen.getByLabelText(
-        'specification.addNeedsRef',
-      )
-      fireEvent.change(needsReferenceSelect, { target: { value: 'new' } })
-      expect(
-        screen.getByLabelText('specification.addNeedsRefTextLabel'),
-      ).toBeInTheDocument()
-      fireEvent.change(needsReferenceSelect, { target: { value: 'none' } })
-      expect(
-        screen.queryByLabelText('specification.addNeedsRefTextLabel'),
-      ).toBeNull()
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.confirmAdd' }),
-      )
-
-      await waitFor(() => {
-        const postCall = fetchMock.mock.calls.find(
-          ([url, init]) =>
-            url === specificationApiPath('/items') &&
-            (init as RequestInit | undefined)?.method === 'POST',
-        )
-        expect(
-          JSON.parse(String((postCall?.[1] as RequestInit | undefined)?.body)),
-        ).toEqual({ requirementIds: [202] })
-      })
-    })
-
-    it('adds selected requirements with a newly described needs reference', async () => {
-      renderRequirementsSpecificationDetailClient()
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-      fireEvent.change(screen.getByLabelText('specification.addNeedsRef'), {
-        target: { value: 'new' },
-      })
-      fireEvent.change(
-        screen.getByLabelText('specification.addNeedsRefTextLabel'),
-        { target: { value: '  IAM-99  ' } },
-      )
-      fireEvent.change(
-        screen.getByLabelText('specification.needsReferenceDescription'),
-        { target: { value: '  Procurement decision  ' } },
-      )
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.confirmAdd' }),
-      )
-
-      await waitFor(() => {
-        const postCall = fetchMock.mock.calls.find(
-          ([url, init]) =>
-            url === specificationApiPath('/items') &&
-            (init as RequestInit | undefined)?.method === 'POST',
-        )
-        expect(
-          JSON.parse(String((postCall?.[1] as RequestInit | undefined)?.body)),
-        ).toEqual({
-          needsReferenceDescription: 'Procurement decision',
-          needsReferenceText: 'IAM-99',
-          requirementIds: [202],
-        })
-      })
-    })
-
-    it('closes the add dialog when Escape is pressed inside the panel', async () => {
-      renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-
-      await screen.findByRole('dialog')
-      fireEvent.change(screen.getByLabelText('specification.addNeedsRef'), {
-        target: { value: 'new' },
-      })
-
-      fireEvent.keyDown(
-        screen.getByLabelText('specification.addNeedsRefTextLabel'),
-        {
-          key: 'Escape',
-        },
-      )
-
-      await waitFor(() => {
-        expect(screen.queryByRole('dialog')).not.toBeInTheDocument()
-      })
-    })
-
-    it('disables needs-reference inputs and help toggles while add is submitting', async () => {
-      renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-
-      await screen.findByRole('dialog')
-      fireEvent.change(screen.getByLabelText('specification.addNeedsRef'), {
-        target: { value: 'new' },
-      })
-
-      const select = screen.getByLabelText('specification.addNeedsRef')
-      const textarea = screen.getByLabelText(
-        'specification.addNeedsRefTextLabel',
-      )
-      const needsRefHelpButton = screen.getByRole('button', {
-        name: 'common.help: specification.addNeedsRef',
-      })
-      const needsRefTextHelpButton = screen.getByRole('button', {
-        name: 'common.help: specification.addNeedsRefTextLabel',
-      })
-
-      let resolvePost:
-        | ((value: { json: () => Promise<unknown>; ok: boolean }) => void)
-        | undefined
-      fetchMock.mockImplementationOnce(
-        () =>
-          new Promise(resolve => {
-            resolvePost = resolve
-          }),
-      )
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.confirmAdd' }),
-      )
-
-      await waitFor(() => {
-        expect(select).toBeDisabled()
-        expect(textarea).toBeDisabled()
-        expect(needsRefHelpButton).toBeDisabled()
-        expect(needsRefTextHelpButton).toBeDisabled()
-      })
-      fireEvent.keyDown(screen.getByRole('dialog'), { key: 'Escape' })
-      expect(screen.getByRole('dialog')).toBeInTheDocument()
-
-      await act(async () => {
-        resolvePost?.({
-          json: async () => ({ ok: true }),
-          ok: true,
-        })
-      })
-    })
-
-    it('keeps the add dialog open when a post-add refresh fails', async () => {
-      renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-
-      const dialog = await screen.findByRole('dialog')
-      failNextSpecificationItemsFetch = true
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.confirmAdd' }),
-      )
-
-      expect(await screen.findByText('common.error')).toHaveAttribute(
-        'role',
-        'alert',
-      )
-      expect(dialog).toBeInTheDocument()
-    })
-
-    it('updates the added-requirements notice when the left filters reveal the requirement', async () => {
-      const addedItem = {
-        ...initialSpecificationItem,
-        id: initialAvailableRequirement.id,
-        itemRef: 'lib:202',
-        specificationItemId: 202,
-        uniqueId: initialAvailableRequirement.uniqueId,
-      }
-      specificationItemsGetHandler = async url => {
-        const matchesAddedRequirement =
-          searchParamsFromPath(url).get('uniqueIdSearch') ===
-          initialAvailableRequirement.uniqueId
-        const items = matchesAddedRequirement ? [addedItem] : []
-        return okJson({
-          items,
-          pagination: {
-            count: items.length,
-            hasMore: false,
-            limit: 50,
-            nextCursor: null,
-          },
-        })
-      }
-
-      renderRequirementsSpecificationDetailClient()
-      await waitForInitialAvailableRequirementsRefresh()
-
-      fireEvent.change(
-        screen.getByRole('textbox', { name: 'search-items-requirements' }),
-        { target: { value: 'DOES-NOT-MATCH' } },
-      )
-      await waitFor(() => {
-        expect(latestItemsTableProps().filterValues).toEqual({
-          uniqueIdSearch: 'DOES-NOT-MATCH',
-        })
-      })
-
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-      await screen.findByRole('dialog')
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.confirmAdd' }),
-      )
-
-      await waitFor(() => {
-        expect(screen.getByRole('status')).toHaveTextContent(
-          'specification.requirementsAddedHiddenByFilters',
-        )
-      })
-      expect(latestItemsTableProps().filterValues).toEqual({
-        uniqueIdSearch: 'DOES-NOT-MATCH',
-      })
-
-      fireEvent.change(
-        screen.getByRole('textbox', { name: 'search-items-requirements' }),
-        { target: { value: initialAvailableRequirement.uniqueId } },
-      )
-
-      await waitFor(() => {
-        expect(screen.getByRole('status')).toHaveTextContent(
-          /^specification\.requirementsAdded$/,
-        )
-      })
-      expect(screen.getByRole('status')).not.toHaveTextContent(
-        'specification.requirementsAddedHiddenByFilters',
-      )
-      expect(latestItemsTableProps().rows).toEqual([addedItem])
-    })
-
-    it('does not announce a matching added requirement as hidden when it is on a later page', async () => {
-      const firstMatchingItem = {
-        ...initialSpecificationItem,
-        id: 303,
-        itemRef: 'lib:303',
-        specificationItemId: 303,
-        uniqueId: 'IAM0001',
-      }
-      const addedMatchingItem = {
-        ...initialSpecificationItem,
-        id: initialAvailableRequirement.id,
-        itemRef: 'lib:202',
-        specificationItemId: 202,
-        uniqueId: initialAvailableRequirement.uniqueId,
-      }
-      specificationItemsGetHandler = async url => {
-        const params = searchParamsFromPath(url)
-        const isMatchProbe = params
-          .getAll('probeRequirementIds')
-          .includes(String(initialAvailableRequirement.id))
-        const items = isMatchProbe ? [addedMatchingItem] : [firstMatchingItem]
-        return okJson({
-          items,
-          pagination: {
-            count: items.length,
-            hasMore: !isMatchProbe,
-            limit: Number(params.get('limit') ?? 50),
-            nextCursor: isMatchProbe ? null : 'later-page',
-          },
-        })
-      }
-
-      renderRequirementsSpecificationDetailClient()
-      await waitForInitialAvailableRequirementsRefresh()
-
-      fireEvent.change(
-        screen.getByRole('textbox', { name: 'search-items-requirements' }),
-        { target: { value: 'IAM' } },
-      )
-      await waitFor(() => {
-        expect(latestItemsTableProps().rows).toEqual([firstMatchingItem])
-      })
-
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      )
-      await screen.findByRole('dialog')
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.confirmAdd' }),
-      )
-
-      await waitFor(() => {
-        expect(screen.getByRole('status')).toHaveTextContent(
-          'specification.requirementsAdded',
-        )
-      })
-      expect(screen.getByRole('status')).not.toHaveTextContent(
-        'specification.requirementsAddedHiddenByFilters',
-      )
-      expect(
-        fetchMock.mock.calls.some(([input]) => {
-          const url = typeof input === 'string' ? input : input.url
-          return (
-            url.startsWith(`${specificationApiPath('/items')}?`) &&
-            searchParamsFromPath(url).getAll('probeRequirementIds')[0] ===
-              String(initialAvailableRequirement.id)
-          )
-        }),
-      ).toBe(true)
-    })
-
-    it('shows a warning when loading more available requirements fails', async () => {
-      failNextAvailableRequirementsFetch = true
-
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        availableRequirements: {
-          hasMore: true,
-          nextCursor: 'cursor-1',
-          rows: [initialAvailableRequirement],
-        },
-      })
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      fireEvent.click(
-        screen.getByRole('button', { name: 'load-more-available' }),
-      )
-
-      expect(
-        await screen.findByText(
-          'specification.loadAvailableRequirementsFailed',
-        ),
-      ).toHaveAttribute('role', 'status')
-    })
-
-    it('opens the specification-local requirement dialog directly from the left-panel create action', async () => {
-      renderRequirementsSpecificationDetailClient()
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 1,
-            name: 'Authorization and IAM',
-          }),
-        ).toBeInTheDocument()
-      })
-
-      const createLocalRequirementButton = screen.getByRole('button', {
-        name: 'specification.newLocalRequirement',
-      })
-      expect(createLocalRequirementButton).toHaveAttribute(
-        'data-developer-mode-name',
-        'table action',
-      )
-      expect(createLocalRequirementButton).toHaveAttribute(
-        'data-developer-mode-context',
-        'requirements specification detail',
-      )
-      expect(createLocalRequirementButton).toHaveAttribute(
-        'data-developer-mode-value',
-        'new local requirement',
-      )
-
-      await act(async () => {
-        fireEvent.click(createLocalRequirementButton)
-      })
-
-      await waitFor(() => {
-        expect(
-          screen.getByRole('heading', {
-            level: 2,
-            name: 'specification.newLocalRequirement',
-          }),
-        ).toBeInTheDocument()
-      })
-      const dialog = screen.getByRole('dialog')
-      expect(within(dialog).queryByLabelText('requirement.area')).toBeNull()
-      expect(
-        within(dialog).queryByText('requirement.requirementPackage'),
-      ).toBeNull()
-
-      const normReferenceFieldset = within(dialog)
-        .getByText('requirement.normReferences')
-        .closest('fieldset')
-      const sidebarGrid = normReferenceFieldset?.parentElement
-      expect(sidebarGrid).toHaveClass('lg:w-full')
-      expect(sidebarGrid?.parentElement).toHaveClass(
-        'lg:grid-cols-[minmax(0,1fr)_minmax(20rem,22rem)]',
-      )
-    })
-
-    it('creates a specification-local requirement and refreshes the application list', async () => {
-      renderRequirementsSpecificationDetailClient()
-      const initialItemsGetCount = fetchMock.mock.calls.filter(
-        ([url, init]) =>
-          String(url).startsWith(`${specificationApiPath('/items')}?`) &&
-          (init as RequestInit | undefined)?.method !== 'POST',
-      ).length
-
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.newLocalRequirement',
-        }),
-      )
-      const dialog = await screen.findByRole('dialog')
-      fireEvent.change(
-        within(dialog).getByRole('textbox', {
-          name: /requirement\.description/,
-        }),
-        { target: { value: 'Local requirement' } },
-      )
-      fireEvent.click(
-        within(dialog).getByRole('button', { name: 'common.save' }),
-      )
-
-      await waitFor(() => {
-        const postCall = fetchMock.mock.calls.find(
-          ([url, init]) =>
-            url === specificationApiPath('/local-requirements') &&
-            (init as RequestInit | undefined)?.method === 'POST',
-        )
-        expect(
-          JSON.parse(String((postCall?.[1] as RequestInit | undefined)?.body)),
-        ).toEqual(expect.objectContaining({ description: 'Local requirement' }))
-      })
-      await waitFor(() => expect(screen.queryByRole('dialog')).toBeNull())
-      await waitFor(() => {
-        const itemsGetCount = fetchMock.mock.calls.filter(
-          ([url, init]) =>
-            String(url).startsWith(`${specificationApiPath('/items')}?`) &&
-            (init as RequestInit | undefined)?.method !== 'POST',
-        ).length
-        expect(itemsGetCount).toBeGreaterThan(initialItemsGetCount)
-        expect(
-          screen.getByRole('table', { name: 'items requirements' }),
-        ).toHaveTextContent('local:401')
-      })
-    })
-
-    it('keeps the local-requirement dialog open when creation fails', async () => {
-      localRequirementPostOk = false
-      renderRequirementsSpecificationDetailClient()
-
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.newLocalRequirement',
-        }),
-      )
-      const dialog = await screen.findByRole('dialog')
-      fireEvent.change(
-        within(dialog).getByRole('textbox', {
-          name: /requirement\.description/,
-        }),
-        { target: { value: 'Local requirement' } },
-      )
-      fireEvent.click(
-        within(dialog).getByRole('button', { name: 'common.save' }),
-      )
-
-      expect(await within(dialog).findByRole('alert')).toHaveTextContent(
-        'Local create failed',
-      )
-    })
-
-    it('keeps dirty local-requirement edits when discard is cancelled', async () => {
-      renderRequirementsSpecificationDetailClient()
-
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.newLocalRequirement',
-        }),
-      )
-      const dialog = await screen.findByRole('dialog')
-      fireEvent.change(
-        within(dialog).getByRole('textbox', {
-          name: /requirement\.description/,
-        }),
-        { target: { value: 'Unsaved local requirement' } },
-      )
-      fireEvent.click(
-        within(dialog).getByRole('button', { name: 'common.close' }),
-      )
-
-      const confirmation = await screen.findByRole('alertdialog')
-      fireEvent.click(
-        within(confirmation).getByRole('button', { name: 'common.cancel' }),
-      )
-      expect(dialog).toBeInTheDocument()
-      expect(
-        within(dialog).getByRole('textbox', {
-          name: /requirement\.description/,
-        }),
-      ).toHaveValue('Unsaved local requirement')
-    })
-  })
+  registerAddingLocalTests(workflowContext)
 
   registerNeedsReferenceTests(workflowContext)
 
@@ -3985,345 +1567,5 @@ describe('RequirementsSpecificationDetailClient', () => {
 
   registerPaginationTests(workflowContext)
 
-  describe('refresh failures and focused regression coverage', () => {
-    it('clears an active available selection when selection answers change', async () => {
-      availableRequirementsSelectionFilter = {
-        applied: false,
-        hasCurrentAnswers: true,
-        hasRequirementSelection: true,
-        hasNoRequirementSelection: false,
-        requirementIds: [202],
-      }
-      renderRequirementsSpecificationDetailClient()
-      const toggle = await screen.findByRole('switch', {
-        name: 'specification.filterWithRequirementSelectionQuestions',
-      })
-      fireEvent.click(toggle)
-      await waitFor(() => expect(toggle).toBeChecked())
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-202' }))
-      expect(
-        screen.getByRole('button', {
-          name: 'specification.addSelectedToSpecification',
-        }),
-      ).toBeInTheDocument()
-
-      fireEvent.click(
-        screen.getByRole('tab', {
-          name: 'specification.requirementSelectionQuestions',
-        }),
-      )
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'notify selection questions changed',
-        }),
-      )
-      fireEvent.click(
-        screen.getByRole('tab', {
-          name: 'specification.availableRequirements',
-        }),
-      )
-
-      await waitFor(() => {
-        expect(
-          screen.queryByRole('button', {
-            name: 'specification.addSelectedToSpecification',
-          }),
-        ).toBeNull()
-      })
-    })
-
-    it('keeps list rows usable when refresh resources reject with Error values', async () => {
-      availableRequirementsGetHandler = async () => {
-        throw new Error('Available requirements offline')
-      }
-      normReferencesGetHandler = async () => {
-        throw new Error('Norm references offline')
-      }
-      specificationItemsGetHandler = async () => {
-        throw 'Items offline'
-      }
-      renderRequirementsSpecificationDetailClient()
-
-      expect(
-        await screen.findByText('Available requirements offline'),
-      ).toHaveAttribute('role', 'status')
-      fireEvent.click(
-        screen.getByRole('button', { name: 'toggle-status-filter-items' }),
-      )
-      await waitFor(() => {
-        expect(
-          fetchMock.mock.calls.some(([input]) =>
-            String(input).startsWith('/api/norm-references'),
-          ),
-        ).toBe(true)
-      })
-      fireEvent.click(
-        screen.getByRole('button', { name: 'sort-description-items' }),
-      )
-      await waitFor(() => {
-        expect(screen.getByRole('alert')).toHaveTextContent(
-          'specification.loadSpecificationItemsFailed',
-        )
-      })
-      expect(
-        screen.getByRole('table', { name: 'items requirements' }),
-      ).toHaveTextContent('lib:31')
-    })
-
-    it('shows the original specification after a non-Error metadata refresh failure', async () => {
-      specificationMetaGetHandler = async () => {
-        throw 'Metadata offline'
-      }
-      renderRequirementsSpecificationDetailClient()
-      fireEvent.click(
-        screen.getByRole('button', { name: 'specification.editSpecification' }),
-      )
-      const dialog = await screen.findByRole('dialog', {
-        name: 'specification.editSpecification',
-      })
-      fireEvent.change(
-        within(dialog).getByRole('textbox', { name: /specification\.name/ }),
-        { target: { value: 'Specification awaiting refresh' } },
-      )
-      fireEvent.click(
-        within(dialog).getByRole('button', { name: /common\.save/i }),
-      )
-
-      expect(
-        await screen.findByText('specification.loadSpecificationFailed'),
-      ).toHaveAttribute('role', 'status')
-      expect(
-        screen.getByRole('heading', { level: 1, name: initialSpec.name }),
-      ).toBeInTheDocument()
-    })
-
-    it('loads needs references after an Error-valued preload failure', async () => {
-      needsReferencesGetHandler = async () => {
-        throw new Error('Needs register offline')
-      }
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        errors: [
-          {
-            key: SPECIFICATION_PRELOAD_ERROR_KEYS.needsReferences,
-            message: 'Needs references missing from preload',
-          },
-        ],
-      })
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: 'specification.newLocalRequirement',
-        }),
-      )
-
-      await waitFor(() => {
-        expect(
-          fetchMock.mock.calls.some(
-            ([input]) =>
-              String(input) === specificationApiPath('/needs-references'),
-          ),
-        ).toBe(true)
-      })
-      expect(screen.getByRole('status')).toHaveTextContent(
-        'specification.partialDataLoadWarning',
-      )
-      expect(screen.getByRole('dialog')).toBeInTheDocument()
-    })
-
-    it('keeps an item without a stable reference visible but unselected', async () => {
-      const itemWithoutRef = {
-        ...initialSpecificationItem,
-        itemRef: undefined,
-      }
-      specificationItemsGetItems = [itemWithoutRef]
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        specificationItems: createSpecificationItemsPage([itemWithoutRef]),
-      })
-
-      expect(
-        screen.getByRole('table', { name: 'items requirements' }),
-      ).toHaveTextContent('101')
-      fireEvent.click(screen.getByRole('button', { name: 'select-row-101' }))
-      expect(
-        screen.queryByRole('button', {
-          name: 'specification.removeSelectedFromSpecification',
-        }),
-      ).toBeNull()
-      await waitForInitialAvailableRequirementsRefresh()
-    })
-
-    it('renders English needs-reference fallbacks for sparse usage metadata', async () => {
-      specificationItemsGetHandler = async () =>
-        okJson({
-          items: [
-            {
-              ...initialSpecificationItem,
-              itemRef: 'lib:english-fallback',
-              specificationItemStatusNameEn: undefined,
-              uniqueId: 'USAGE-ENGLISH-FALLBACK',
-              version: {
-                ...initialSpecificationItem.version,
-                description: undefined,
-                typeNameEn: undefined,
-              },
-            },
-          ],
-        })
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        availableNeedsRefs: [
-          {
-            createdAt: '',
-            description: 'Sparse usage',
-            id: 82,
-            libraryItemCount: 1,
-            linkedItemCount: 1,
-            specificationLocalRequirementCount: 0,
-            text: 'NEED-82',
-            updatedAt: '',
-          },
-        ],
-      })
-      fireEvent.click(
-        screen.getByRole('tab', { name: /specification\.needsReferences/ }),
-      )
-      fireEvent.click(
-        screen.getByRole('button', {
-          name: /specification\.toggleNeedsReferenceUsage/,
-        }),
-      )
-
-      expect(
-        await screen.findByText('USAGE-ENGLISH-FALLBACK'),
-      ).toBeInTheDocument()
-      expect(screen.getAllByText('—')).toHaveLength(3)
-    })
-
-    it('offers the management report for a specification in management', async () => {
-      renderRequirementsSpecificationDetailClient({
-        ...createInitialData(),
-        spec: {
-          ...initialSpec,
-          lifecycleStatus: {
-            id: 4,
-            nameEn: 'Management',
-            nameSv: 'Förvaltning',
-          },
-          specificationLifecycleStatusId: 4,
-        },
-      })
-      await waitForInitialAvailableRequirementsRefresh()
-      const moreActions = (
-        latestItemsTableProps().floatingActions as Array<{
-          id: string
-          menuItems?: Array<{
-            id: string
-            onClick?: (target?: HTMLButtonElement | null) => void
-          }>
-        }>
-      ).find(action => action.id === 'more-actions')
-      const managementReport = screen.getByRole('menuitem', {
-        name: 'specification.downloadProfileReportPdf.specification.reportProfiles.management',
-      })
-      fireEvent.click(managementReport)
-
-      expect(moreActions?.menuItems?.map(item => item.id)).toContain(
-        'pdf-management',
-      )
-      expect(pdfDownloadState.download).toHaveBeenCalledWith(
-        expect.objectContaining({
-          restoreFocusTo: managementReport,
-          url: '/en/specifications/8/reports/pdf/management',
-        }),
-      )
-    })
-
-    it.each([
-      ['Error', new Error('Available continuation offline')],
-      ['non-Error', 'Available continuation offline'],
-    ])(
-      'keeps available rows after an %s load-more failure',
-      async (_kind, failure) => {
-        availableRequirementsGetHandler = async url => {
-          if (searchParamsFromPath(url).has('cursor')) throw failure
-          return okJson({
-            pagination: { hasMore: true, nextCursor: 'cursor-2' },
-            requirements: [initialAvailableRequirement],
-          })
-        }
-        renderRequirementsSpecificationDetailClient({
-          ...createInitialData(),
-          availableRequirements: {
-            hasMore: true,
-            nextCursor: 'cursor-2',
-            rows: [initialAvailableRequirement],
-          },
-        })
-        const loadMore = await screen.findByRole('button', {
-          name: 'load-more-available',
-        })
-        fireEvent.click(loadMore)
-
-        expect(
-          await screen.findByText(
-            failure instanceof Error
-              ? failure.message
-              : 'specification.loadAvailableRequirementsFailed',
-          ),
-        ).toHaveAttribute('role', 'status')
-        expect(
-          screen.getByRole('table', { name: 'available requirements' }),
-        ).toHaveTextContent('202')
-      },
-    )
-
-    it('keeps selection filtering on while an empty continuation finishes the list', async () => {
-      availableRequirementsSelectionFilter = {
-        applied: false,
-        hasCurrentAnswers: true,
-        hasRequirementSelection: true,
-        hasNoRequirementSelection: false,
-        requirementIds: [202],
-      }
-      availableRequirementsGetHandler = async url => {
-        if (searchParamsFromPath(url).has('cursor')) return okJson({})
-        return okJson({
-          pagination: { hasMore: true, nextCursor: 'filtered-cursor' },
-          requirements: [initialAvailableRequirement],
-          selectionFilter: {
-            ...availableRequirementsSelectionFilter,
-            applied: url.includes('applyRequirementSelectionFilter=true'),
-          },
-        })
-      }
-      renderRequirementsSpecificationDetailClient()
-      const toggle = await screen.findByRole('switch', {
-        name: 'specification.filterWithRequirementSelectionQuestions',
-      })
-      fireEvent.click(toggle)
-      const loadMore = await screen.findByRole('button', {
-        name: 'load-more-available',
-      })
-      fireEvent.click(loadMore)
-
-      await waitFor(() => {
-        expect(
-          availableRequirementsFetchUrls().some(url => {
-            const params = searchParamsFromPath(url)
-            return (
-              params.get('cursor') === 'filtered-cursor' &&
-              params.get('applyRequirementSelectionFilter') === 'true'
-            )
-          }),
-        ).toBe(true)
-        expect(
-          screen.queryByRole('button', { name: 'load-more-available' }),
-        ).toBeNull()
-      })
-      expect(
-        screen.getByRole('table', { name: 'available requirements' }),
-      ).toHaveTextContent('202')
-    })
-  })
+  registerResilienceTests(workflowContext)
 })
